% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{addMessage}
\alias{addMessage}
\title{add an message to list of messages for printing #}
\usage{
addMessage(messageList, message)
}
\description{
add an message to list of messages for printing #
}
\keyword{internal}
