% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralPreprocessing.R
\name{addPreprocessingStep}
\alias{addPreprocessingStep}
\title{function to use by preprocessing step}
\usage{
addPreprocessingStep(object, newSpectra, preprocessingInfo)
}
\arguments{
\item{object}{\code{\link{SpectraInTime-class}}}

\item{newSpectra}{numeric matrix new spectral data}

\item{preprocessingInfo}{list of preprocessing info}
}
\value{
\code{\link{SpectraInTime-class}}
}
\description{
change the spectra and add preprocessing info, check whehther the object is valid
}
\keyword{internal}
