% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/objectLinking.R
\name{checkCompatible}
\alias{checkCompatible}
\alias{checkCompatible,SpectraInTime,ProcessTimes-method}
\alias{[ProcessTimes,SpectraInTime-method}
\alias{[SpectraInTime,ProcessTimes-method}
\alias{checkCompatible,ProcessTimes,SpectraInTime-method}
\title{Check object compatibility}
\usage{
checkCompatible(x, y, ...)

\S4method{checkCompatible}{SpectraInTime,ProcessTimes}(x, y)

\S4method{checkCompatible}{ProcessTimes,SpectraInTime}(x, y)
}
\arguments{
\item{x}{first object}

\item{y}{second object}

\item{...}{additional parameters}
}
\value{
no output, produces an error when object are not compatible with each other

\code{TRUE} when the aer competible, otherwise it stops and prints a list of error messages
}
\description{
Check wheter 2 objects are compatible before using them together e.g. in time aligment using a time file with matching experiment name.
}
