% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralNMF.R
\name{checkForRedundantSources}
\alias{checkForRedundantSources}
\title{Check for redunt NMF source vectors}
\usage{
checkForRedundantSources(seed)
}
\arguments{
\item{seed}{nmfModel object containing initialization of the factor matrices}
}
\value{
boolean vector, indicating which source vector(s) are redundant
}
\description{
Check if any of the source vectors in the initialized NMF model are redundant
 and should be omitted from the actual NMF analysis
}
\author{
Nicolas Sauwen
}
