% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{checkIdenticalClass}
\alias{checkIdenticalClass}
\title{check wether all elements of of the same class}
\usage{
checkIdenticalClass(listOfObjects, class)
}
\arguments{
\item{listOfObjects}{a list of S4 objects to check}

\item{class}{a class to compare with}
}
\value{
logical value \code{TRUE} if all objects are of the correct class
}
\description{
check wether all elements of of the same class
}
\author{
Adriaan Blommaert
}
