% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralPreprocessing.R
\name{checkMethod}
\alias{checkMethod}
\title{check method in list of allowd method}
\usage{
checkMethod(method, allowedMethods)
}
\arguments{
\item{method}{character of method to check}

\item{allowedMethods}{character vector of allowed method}
}
\value{
nothing just stop if method not allowed
}
\description{
check method in list of allowd method
}
\keyword{internal}
