% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{checkTimeUnit}
\alias{checkTimeUnit}
\title{check timeeUnit  and produce error if not correct unit}
\usage{
checkTimeUnit(timeUnit)
}
\arguments{
\item{timeUnit}{character vector}
}
\description{
check timeeUnit  and produce error if not correct unit
}
\keyword{internal}
