% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSpectralObjects.R
\name{combineSpectralObjects}
\alias{combineSpectralObjects}
\title{Function to combine \code{\link{SpectraInTime-class}} objects containing 1 spectrum each}
\usage{
combineSpectralObjects(objectList, timeRange, checkNames = TRUE)
}
\arguments{
\item{objectList}{List of \code{\link{SpectraInTime-class}} objects to be combined}

\item{timeRange}{Numeric value, equal to the maximum time of the measured spectra.}

\item{checkNames}{Boolean - if \code{TRUE}, the experiment name of the spectral objects will be compared to see if these spectral objects belong to the same experiment}
}
\value{
\code{\link{SpectraInTime-class}}
}
\description{
Function to combine \code{\link{SpectraInTime-class}} objects containing 1 spectrum each
}
\author{
Nicolas Sauwen
}
