% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralNMF.R
\name{computeNMFResidu}
\alias{computeNMFResidu}
\title{NMF relative residual per observation}
\usage{
computeNMFResidu(object, NMFResult)
}
\arguments{
\item{object}{\code{\link{SpectraInTime-class}}}

\item{NMFResult}{Fitted NMF model}
}
\value{
Dataframe, containing time (observation) vector and residual vector
}
\description{
Compute relative residual per observation of an NMF fit to a spectral data set
}
\author{
nsauwen
}
