% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{convertListToS4}
\alias{convertListToS4}
\title{convert an a named list to S4 object}
\usage{
convertListToS4(list, class)
}
\arguments{
\item{list}{list}

\item{class}{character vector indicating S4 class}
}
\description{
convert an a named list to S4 object
}
\keyword{internal}
