% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{convertS4ToList}
\alias{convertS4ToList}
\title{convert an S4 object into a list}
\usage{
convertS4ToList(object)
}
\arguments{
\item{object}{and S4 object}
}
\description{
convert an S4 object into a list
}
\keyword{internal}
