% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{e}
\alias{e}
\title{Create an \code{\link{ElementsToSelect-class}} from a numeric vector or multiple numeric values or vectors}
\usage{
e(x, ...)
}
\arguments{
\item{x}{numeric  vector}

\item{...}{additional numeric vectors}
}
\value{
\code{\link{ElementsToSelect-class}} with unique elements
}
\description{
Create an \code{\link{ElementsToSelect-class}} from a numeric vector or multiple numeric values or vectors
}
\examples{
e( 1 , 5, 4.5  )
e( 1:10 , c(4 , 5 , 6 ) , 7 )
}
