% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\docType{class}
\name{ElementsToSelect-class}
\alias{ElementsToSelect-class}
\title{Elements S4 class useful for closest elements subsetting}
\description{
Elements S4 class useful for closest elements subsetting
}
\section{Slots}{

\describe{
\item{\code{elements}}{numeric vector of elements}
}}

\author{
Adriaan Blommaert
}
