% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/objectSpectraInTime.R
\name{firstSpectrum}
\alias{firstSpectrum}
\alias{firstSpectrum,SpectraInTime-method}
\alias{firstSpectrum,numeric-method}
\title{Get the first spectrum}
\usage{
firstSpectrum(object, ...)

\S4method{firstSpectrum}{SpectraInTime}(object)

\S4method{firstSpectrum}{numeric}(object)
}
\arguments{
\item{object}{S4 object}

\item{...}{additional parameters}
}
\value{
numeric vector containing observations first spectrum
}
\description{
Get the first spectrum
}
