% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{flagVectorInInterval}
\alias{flagVectorInInterval}
\title{flag vector elements within interval (closed interval )}
\usage{
flagVectorInInterval(vector, interval)
}
\arguments{
\item{vector}{numeric vector}

\item{interval}{numeric vecor with 2 values}
}
\description{
flag vector elements within interval (closed interval )
}
\note{
if interval is \code{NULL} then no selection  is performed
}
\author{
Adriaan Blommaert
}
\keyword{internal}
