% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{getClosestElements}
\alias{getClosestElements}
\title{get closest element for elements looked up in a vector}
\usage{
getClosestElements(vector, elements)
}
\arguments{
\item{vector}{numeric vector values to match to}

\item{elements}{numeric vector}
}
\description{
get closest element for elements looked up in a vector
}
\author{
Adriaan Blommaert
}
\keyword{internal}
