% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectSpectraInTime.R
\name{getDefaultSumFunc}
\alias{getDefaultSumFunc}
\title{function to get default summary functions}
\usage{
getDefaultSumFunc()
}
\value{
character vector of functions
}
\description{
function to get default summary functions
}
