% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{getDefaultTimeFormat}
\alias{getDefaultTimeFormat}
\title{function to get default time format in the package}
\usage{
getDefaultTimeFormat()
}
\value{
character vector specifiying a time format

character string iwth default time format
}
\description{
function to get default time format in the package
}
