% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R
\docType{methods}
\name{getDimensionReduction}
\alias{getDimensionReduction}
\title{generic function to extract \code{dimensionReduction}-slot}
\usage{
getDimensionReduction(object, ...)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\value{
dimension reduction slot of an object
}
\description{
generic function to extract \code{dimensionReduction}-slot
}
