% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/subsetting.R
\docType{methods}
\name{getElements}
\alias{getElements}
\alias{getElements,ElementsToSelect-method}
\title{generic function to extract \code{elements}-slot}
\usage{
getElements(object, ...)

\S4method{getElements}{ElementsToSelect}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function to extract \code{elements}-slot
}
