% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/objectSpectraInTime.R
\docType{methods}
\name{getExperimentName}
\alias{getExperimentName}
\alias{getExperimentName,SpectraInTime-method}
\title{generic function to extract \code{experimentName}-slot}
\usage{
getExperimentName(object, ...)

\S4method{getExperimentName}{SpectraInTime}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\value{
string containing experiment name
}
\description{
generic function to extract \code{experimentName}-slot
}
