% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/objectSpectraInTime.R
\docType{methods}
\name{getExtraInfo}
\alias{getExtraInfo}
\alias{getExtraInfo,SpectraInTime-method}
\title{generic function to extract \code{extraInfo}-slot}
\usage{
getExtraInfo(object, ...)

\S4method{getExtraInfo}{SpectraInTime}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\value{
list of extraInfo
}
\description{
generic function to extract \code{extraInfo}-slot
}
