% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectProcessTimes.R
\name{getListOfSpectraExample}
\alias{getListOfSpectraExample}
\title{get example list of spectra}
\usage{
getListOfSpectraExample()
}
\value{
list of \code{\link{SpectraInTime-class}}
}
\description{
get example list of spectra
}
