% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/objectSpectraInTime.R
\docType{methods}
\name{getPreprocessing}
\alias{getPreprocessing}
\alias{getPreprocessing,SpectraInTime-method}
\title{generic function to extract \code{preprocessing}-slot}
\usage{
getPreprocessing(object, ...)

\S4method{getPreprocessing}{SpectraInTime}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\value{
list with preprocessing steps
}
\description{
generic function to extract \code{preprocessing}-slot
}
