% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectProcessTimes.R
\name{getProcessTimesExample}
\alias{getProcessTimesExample}
\title{get a minimal \code{\link{ProcessTimes-class}} example based on 
\code{\link{getSpectraInTimeExample}}}
\usage{
getProcessTimesExample()
}
\value{
\code{\link{ProcessTimes-class}}
}
\description{
get a minimal \code{\link{ProcessTimes-class}} example based on 
\code{\link{getSpectraInTimeExample}}
}
\examples{
 getProcessTimesExample()
}
\author{
Adriaan Blommaert
}
