% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectProcessTimes.R
\name{getProcessTimesFrameExample}
\alias{getProcessTimesFrameExample}
\title{get mimimal example \code{\link{ProcessTimesFrame-class}}}
\usage{
getProcessTimesFrameExample()
}
\value{
\code{\link{ProcessTimes-class}}
}
\description{
get mimimal example \code{\link{ProcessTimesFrame-class}}
}
\author{
Adriaan Blommaert
}
