% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/subsetting.R
\docType{methods}
\name{getRange}
\alias{getRange}
\alias{getRange,RangeToSubset-method}
\title{generic function to extract \code{range}-slot}
\usage{
getRange(object, ...)

\S4method{getRange}{RangeToSubset}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function to extract \code{range}-slot
}
