% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/objectSpectraInTime.R
\docType{methods}
\name{getSpectra}
\alias{getSpectra}
\alias{getSpectra,SpectraInTime-method}
\title{generic function to extract \code{spectra}-slot}
\usage{
getSpectra(object, ...)

\S4method{getSpectra}{SpectraInTime}(object)

\S4method{getSpectra}{SpectraInTime}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\value{
matrix of spectra
}
\description{
generic function to extract \code{spectra}-slot
}
