% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectSpectraInTimeComp.R
\name{getSpectraInTimeCompExample}
\alias{getSpectraInTimeCompExample}
\title{Artificial example of \code{\link{SpectraInTimeComp-class}}}
\usage{
getSpectraInTimeCompExample()
}
\value{
\code{\link{SpectraInTimeComp-class}}
}
\description{
Example \code{\link{SpectraInTime-class}} with nmf result 
using random initialization with rank 2
}
\examples{
\donttest{
	test <-  getSpectraInTimeCompExample() 
}
}
\author{
Adriaan Blommaert
}
