% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectSpectraInTime.R
\name{getSpectraInTimeExample}
\alias{getSpectraInTimeExample}
\title{Artificial example \code{\link{SpectraInTime-class}}}
\usage{
getSpectraInTimeExample(showPlots = FALSE)
}
\arguments{
\item{showPlots}{logical indicator to show plots}
}
\value{
\code{\link{SpectraInTime-class}}
}
\description{
exponential conversion from 2 concentrations with gaussion curves for spectra at different wavelength per compounds
}
\examples{
ex1  <-   getSpectraInTimeExample()
}
\author{
Adriaan Blommaert
}
