% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/objectSpectraInTime.R
\docType{methods}
\name{getSpectralAxis}
\alias{getSpectralAxis}
\alias{getSpectralAxis,SpectraInTime-method}
\title{generic function to extract \code{spectralAxis}-slot}
\usage{
getSpectralAxis(object, ...)

\S4method{getSpectralAxis}{SpectraInTime}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\value{
numeric vector containing wavelengths
}
\description{
generic function to extract \code{spectralAxis}-slot
}
