% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/objectSpectraInTime.R
\docType{methods}
\name{getStartTime}
\alias{getStartTime}
\alias{getStartTime,SpectraInTime-method}
\title{generic function to extract \code{startTime}-slot}
\usage{
getStartTime(object, ...)

\S4method{getStartTime}{SpectraInTime}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\value{
character vector with start time
}
\description{
generic function to extract \code{startTime}-slot
}
