% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{getTimeConversionFactor}
\alias{getTimeConversionFactor}
\title{transform timeUnit to convert from seconds to specified unit}
\usage{
getTimeConversionFactor(timeUnit)
}
\description{
@param timeUnit character vector
}
\keyword{internal}
