% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/objectSpectraInTime.R
\docType{methods}
\name{getUnits}
\alias{getUnits}
\alias{getUnits,SpectraInTime-method}
\title{generic function to extract \code{units}-slot}
\usage{
getUnits(object, ...)

\S4method{getUnits}{SpectraInTime}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\value{
list of units
}
\description{
generic function to extract \code{units}-slot
}
