% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/objectSpectraInTime.R
\docType{methods}
\name{getWavelengths}
\alias{getWavelengths}
\alias{getWavelengths,SpectraInTime-method}
\title{generic function to extract \code{wavelengths}-slot}
\usage{
getWavelengths(object, ...)

\S4method{getWavelengths}{SpectraInTime}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function to extract \code{wavelengths}-slot
}
