% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralNMF.R
\name{includeRedundantSources}
\alias{includeRedundantSources}
\title{Re-introduce redundant sources in NMF-model}
\usage{
includeRedundantSources(NMFResult, seed_orig, redundantSources)
}
\arguments{
\item{NMFResult}{Fitted NMF model}

\item{seed_orig}{Initial NMF model}

\item{redundantSources}{boolean vector, obtained from \code{\link{checkForRedundantSources}}}
}
\value{
Final NMF model with redundant sources re-introduced
}
\description{
Re-introduce redundant source vectors and corresponding zero abundances 
into final NMF result
}
\author{
Nicolas Sauwen
}
