% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralNMF.R
\name{initializeNMFModel}
\alias{initializeNMFModel}
\title{Initialize NMF model with initial spectral data}
\usage{
initializeNMFModel(initSpectralData, spectra, wavelengths = NULL)
}
\arguments{
\item{initSpectralData}{this can be a list of spectralData objects, containing 
the pure component spectra. It can also be either of the NMF factor matrices with initial values}

\item{spectra}{spectral matrix, with wavelengths as its rows and time points as its columns}

\item{wavelengths}{vector of wavelength values}
}
\description{
Initialize NMF model with initial spectral data
}
