% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/objectSpectraInTime.R
\docType{methods}
\name{lastSpectrum}
\alias{lastSpectrum}
\alias{lastSpectrum,numeric-method}
\alias{lastSpectrum,SpectraInTime-method}
\title{Get the last spectrum}
\usage{
lastSpectrum(object, ...)

\S4method{lastSpectrum}{numeric}(object)

\S4method{lastSpectrum}{SpectraInTime}(object)
}
\arguments{
\item{object}{S4 object}

\item{...}{additional parameters}
}
\description{
Get the last spectrum
}
