% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralPreprocessing.R
\name{multiplicativeScatterCorrect}
\alias{multiplicativeScatterCorrect}
\title{multiplicative scatter correction on 1 function}
\usage{
multiplicativeScatterCorrect(x, xRef)
}
\arguments{
\item{x}{numerical vector of measurements}
}
\value{
corrected spectrum
}
\description{
multiplicative scatter correction on 1 function
}
\keyword{internal}
