% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralNMF.R
\name{nonNegativePreprocessing}
\alias{nonNegativePreprocessing}
\title{condition datamatrix to input in and condition properly for NMF}
\usage{
nonNegativePreprocessing(spectra, method = "")
}
\arguments{
\item{spectra}{matrix of spectra}

\item{method}{name of the NMF method to be used.}
}
\value{
matrix, with wavelengths as its rows and time points as its columns
}
\description{
condition datamatrix to input in and condition properly for NMF
}
\details{
put negative values to zero, transpose, an add small value zero row (wavelength with only zeros)
}
