% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R,
%   R/spectralPreprocessing.R
\docType{methods}
\name{preprocess}
\alias{preprocess}
\alias{preprocess,SpectraInTime,list-method}
\alias{preprocess,SpectraInTime,SpectraInTime-method}
\title{generic function to preprocess an S4 object}
\usage{
preprocess(object, with)

\S4method{preprocess}{SpectraInTime,list}(object, with)

\S4method{preprocess}{SpectraInTime,SpectraInTime}(object, with)
}
\arguments{
\item{object}{a S4 class object}

\item{with}{an other object containing preprocessing information: other S4 object, list or expression}
}
\description{
generic function to preprocess an S4 object
}
\examples{

 object1   <-  getSpectraInTimeExample()
 object2   <-  getSpectraInTimeExample

}
