% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralPreprocessing.R
\name{preprocessSpectraInTimeWithList}
\alias{preprocessSpectraInTimeWithList}
\title{internal function to wrap multiple preprocessing steps}
\usage{
preprocessSpectraInTimeWithList(object, with)
}
\description{
internal function to wrap multiple preprocessing steps
}
\keyword{internal}
