% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectProcessTimes.R
\docType{class}
\name{ProcessTimes-class}
\alias{ProcessTimes-class}
\alias{ProcessTimes}
\alias{TemperatureInfo}
\alias{temperatureInfo}
\alias{temperatureinfo}
\title{S4 Class key process times}
\description{
S4 Class key process times
}
\section{Slots}{

\describe{
\item{\code{experimentName}}{character vector with name of the experiment}

\item{\code{timeHeatingAboveMin}}{time when experiment above minimum temperature}

\item{\code{timeStartReaction}}{time start reaction (end of heating ramp)}

\item{\code{timeEndProcess}}{time timeEndProcess  time end of the process, when cooling down starts}

\item{\code{Tset}}{the maximum temperature to indicate \code{timeStartReaction}}

\item{\code{comments}}{character vector of comments when \code{NA} values are produced}
}}

\author{
Adriaan Blommaert
}
