% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectProcessTimes.R
\docType{class}
\name{ProcessTimesFrame-class}
\alias{ProcessTimesFrame-class}
\title{ProcessTimes-class}
\value{
\code{\link{ProcessTimes-class}}
}
\description{
S4 Class key process times in a data frame, every line is convertable to a \code{\link{ProcessTimes-class} }
}
\section{Slots}{

\describe{
\item{\code{processTimes}}{data.frame with every line process times of an experiment}
}}

\author{
Adriaan Blommaert
}
