% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\docType{methods}
\name{r}
\alias{r}
\alias{r,numeric,numeric-method}
\alias{r,RangeToSubset,missing-method}
\title{create a \code{\link{RangeToSubset-class}} object from 2 elements or from a vector}
\usage{
r(x, y)

\S4method{r}{numeric,numeric}(x, y)

\S4method{r}{RangeToSubset,missing}(x, y)
}
\arguments{
\item{x}{numeric value or vector of numeric values}

\item{y}{numeric value missing when x is a vector of values}
}
\description{
create a \code{\link{RangeToSubset-class}} object from 2 elements or from a vector
}
