% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\docType{class}
\name{RangeToSubset-class}
\alias{RangeToSubset-class}
\alias{RangeToSubset}
\alias{rangetosubset}
\alias{Rangetosubset}
\title{RangeToSubset-class}
\description{
Allows for subsetting a range of actual values instead of a range of indicators
}
\section{Slots}{

\describe{
\item{\code{range}}{numeric vector with min and max value}
}}

\author{
Adriaan Blommaert
}
