% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectProcessTimes.R
\name{readProcessTimes}
\alias{readProcessTimes}
\title{read .csv file as process times}
\usage{
readProcessTimes(path, timeFormat = "\%Y-\%m-\%d \%H:\%M:\%OS")
}
\arguments{
\item{path}{to the file containing process times information}

\item{timeFormat}{character specifying time format \code{\link[base]{as.POSIXct}}}
}
\value{
\code{\link{ProcessTimesFrame-class}}

\code{\link{ProcessTimes-class}}
}
\description{
read .csv file as process times
}
\examples{
 readProcessTimes( getPathProcessTimesExample() , timeFormat =  "\%Y-\%m-\%d \%H:\%M:\%S" )
}
