% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSPC.R
\name{readSPC}
\alias{readSPC}
\title{Read-in of a SPC file.}
\usage{
readSPC(filename, keys.log2data = TRUE, keys.hdr2data = FALSE)
}
\arguments{
\item{filename}{Character vector expressing the name of the SPC file (just the name, not the directory).}

\item{keys.log2data}{Logical vector indicating whether the full information (consisting of additional information on the experimental conditions) needs to be parsed from the SPC file or not (TRUE indicates that the full information should be parsed from the SPC file). The default value is FALSE.}

\item{keys.hdr2data}{a character vector of header object to add to backgroundInformation}
}
\value{
\code{\link{SpectraInTime-class}}
}
\description{
This function is an adaptation of the 'read.spc' function of the 'hyperSpec' package :
  Claudia Beleites and Valter Sergo: `hyperSpec: a package to handle
  hyperspectral data sets in R, R package version 0.98-20161118.
  http://hyperspec.r-forge.r-project.org.
}
