% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralNMF.R
\name{removeRedundantSources}
\alias{removeRedundantSources}
\title{Remove redundant sources from the initial NMF model}
\usage{
removeRedundantSources(seed, redundantSources)
}
\arguments{
\item{seed}{nmfModel object containing initialization of the factor matrices}

\item{redundantSources}{boolean vector, obtained from \code{\link{checkForRedundantSources}}}
}
\value{
nmfModel object with redundant sources removed from initial factor matrices
}
\description{
Remove redundant sources from the initial NMF model
}
\author{
Nicolas Sauwen
}
