% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralPreprocessing.R
\name{scatterCorrectSpectra}
\alias{scatterCorrectSpectra}
\title{perform multiplicative scatter correction on spectralData}
\usage{
scatterCorrectSpectra(
  object,
  referenceSpectra = NULL,
  referenceMethod = "mean"
)
}
\description{
perform multiplicative scatter correction on spectralData
}
\author{
Adriaan Blommaert
}
\keyword{internal}
