% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/objectSpectraInTime.R
\docType{methods}
\name{setExperimentName<-}
\alias{setExperimentName<-}
\alias{setExperimentName<-,SpectraInTime-method}
\alias{setTimePointsAlt<-,SpectraInTime-method}
\title{set the experiment name}
\usage{
setExperimentName(object) <- value

\S4method{setExperimentName}{SpectraInTime}(object) <- value

\S4method{setTimePointsAlt}{SpectraInTime}(object) <- value
}
\arguments{
\item{object}{a S4 class object}

\item{value}{a vector of time points}
}
\value{
\code{\link{SpectraInTime-class}} with modified experiment name
}
\description{
set the experiment name
}
