% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R
\docType{methods}
\name{setTimePointsAlt<-}
\alias{setTimePointsAlt<-}
\title{set time alternative time axis}
\usage{
setTimePointsAlt(object) <- value
}
\arguments{
\item{object}{a S4 class object}

\item{value}{a vector of time points}
}
\value{
\code{\link{SpectraInTime-class}} with modified timePointsAlt axis
}
\description{
set time alternative time axis
}
