% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralAnalysis.R
\docType{package}
\name{spectralAnalysis}
\alias{spectralAnalysis}
\alias{spectralAnalysis-package}
\title{spectralAnalysis: a package to read-in, pre-process, visualise and analyse spectral data}
\description{
spectralAnalysis: a package to read-in, pre-process, visualise and analyse spectral data
}
