% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralPreprocessing.R
\name{spectralNormalization}
\alias{spectralNormalization}
\title{normalization function}
\usage{
spectralNormalization(
  object,
  method = "normalize",
  spectralRange = r(-Inf, Inf),
  spectralAxisVal = NULL,
  scaleFunction = "sd",
  meanFunction = NULL
)
}
\description{
normalization function
}
\keyword{internal}
