% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralPLS.R
\name{spectralPLSCalibration}
\alias{spectralPLSCalibration}
\title{Compute PLS model}
\usage{
spectralPLSCalibration(objectList, UPLC_DF, ncomp = 10)
}
\arguments{
\item{objectList}{list of SPC files}

\item{UPLC_DF}{dataframe with UPLC data, which should contain the following columns: experiment, time, and 1 column per compound}

\item{ncomp}{number of PLS components, defaults to  10}
}
\value{
PLS model, as obtained from \code{\link[pls]{plsr}}
}
\description{
Compute PLS model
}
\author{
Nicolas Sauwen
}
