% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralPLS.R
\name{spectralPlsPrediction}
\alias{spectralPlsPrediction}
\title{Perform PLS prediction}
\usage{
spectralPlsPrediction(spectralObject, plsModel, nComp)
}
\arguments{
\item{spectralObject}{\code{\link{SpectraInTime-class}}}

\item{plsModel}{PLS model as obtained from \code{\link{spectralPLSCalibration}}}

\item{nComp}{Number of components}
}
\value{
\code{\link{SpectraInTimeComp-class}} which includes PLS model + prediction
}
\description{
Perform PLS prediction
}
\author{
Nicolas Sauwen
}
