% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{subset.SpectraTime}
\alias{subset.SpectraTime}
\title{internal function for subsetting}
\usage{
\method{subset}{SpectraTime}(x, i, j, timeUnit = "seconds", timePointsAlt = FALSE, drop = "")
}
\description{
internal function for subsetting
}
\keyword{internal}
