% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectSpectraInTime.R
\docType{class}
\name{SummaryByWavelengths-class}
\alias{SummaryByWavelengths-class}
\alias{SummaryByWavelengths}
\alias{plot,SummaryByWavelengths,ANY-method}
\alias{getSpectra,SummaryByWavelengths-method}
\title{S4-class for summary object SpectraInTime}
\usage{
\S4method{plot}{SummaryByWavelengths,ANY}(x, y, colors = NULL)

\S4method{getSpectra}{SummaryByWavelengths}(object)
}
\arguments{
\item{x}{\code{\link{SummaryByWavelengths-class}}}

\item{y}{no used, for internal consistency}

\item{colors}{a vector of colors used in the plot, default to \code{NULL}, when default colors are used}
}
\description{
S4-class for summary object SpectraInTime

plot summary of \code{\link{SpectraInTime-class}}
}
\examples{
summary               <-  summary( getSpectraInTimeExample() )
plot( summary )
}
\keyword{internal}
