% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentFunctions.R
\name{testListOfSpectra}
\alias{testListOfSpectra}
\title{internal function to check all elements are spectraInTime}
\usage{
testListOfSpectra(list)
}
\description{
internal function to check all elements are spectraInTime
}
\keyword{internal}
