% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentFunctions.R
\name{timeAlign.listOfSpectra}
\alias{timeAlign.listOfSpectra}
\title{Align all spectra in a list}
\usage{
timeAlign.listOfSpectra(
  spectralDataList,
  processTimes,
  cutCooling,
  cutBeforeMinTemp
)
}
\description{
Align all spectra in a list
}
\keyword{internal}
