% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentFunctions.R
\name{timeAlign.SpectraInTime}
\alias{timeAlign.SpectraInTime}
\title{align 1 SpectraInTime with 1 cooling information}
\usage{
timeAlign.SpectraInTime(x, y, cutCooling, cutBeforeMinTemp)
}
\arguments{
\item{x}{\code{\link{SpectraInTime-class}}}

\item{y}{\code{\link{ProcessTimes-class}}}
}
\description{
align 1 SpectraInTime with 1 cooling information
}
\keyword{internal}
