% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralIntegration.R
\name{trapz}
\alias{trapz}
\title{Integration via trapezium rule}
\usage{
trapz(x, y)
}
\description{
Originally from caTools package, included to avoid dependency on old package
with correction inverse scale, should give still possitive values (taken absolute values )
}
\keyword{internal}
