% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralNMF.R
\name{upsampleNMFResult}
\alias{upsampleNMFResult}
\title{Upsample NMF result to original temporal resolution}
\usage{
upsampleNMFResult(NMFResult, timePoints, subsamplingFactor, shift = 0)
}
\arguments{
\item{NMFResult}{Fitted NMF model}

\item{timePoints}{Original time points}

\item{subsamplingFactor}{Subsampling factor}

\item{shift}{Integer that correctly shifts subsampling index when 
applying NMF to multiple experiments}
}
\value{
Upsampled NMF model
}
\description{
Upsample NMF result to original temporal resolution
}
\author{
Nicolas Sauwen
}
