% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{changeDevSettings}
\alias{changeDevSettings}
\title{Function to overwrite default settings}
\usage{
changeDevSettings(defaultSettings, replaceSettings)
}
\arguments{
\item{defaultSettings}{list of settings}

\item{replaceSettings}{list of settings to overwrite defaults or add to defaults}
}
\value{
list
}
\description{
Changes are printed print to the console when changing default settings and adding to default settings
}
\keyword{internal}
