% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralNMF.R
\name{getNMFInputMatrix}
\alias{getNMFInputMatrix}
\title{Get spectralData as input NMF model}
\usage{
getNMFInputMatrix(object, method = "")
}
\arguments{
\item{object}{object of the 'spectralData' class, such as a raw SPC file}

\item{method}{name of the NMF method to be used.}
}
\value{
spectral matrix, with wavelengths as its rows and time points as its columns
}
\description{
Extract spectral input matrix from \code{\link{SpectraInTime-class}} and condition properly for NMF modeling
}
\author{
Nicolas Sauwen
}
