% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectProcessTimes.R
\name{getPathProcessTimesExample}
\alias{getPathProcessTimesExample}
\title{example path process times 
ecport}
\usage{
getPathProcessTimesExample()
}
\description{
example path process times 
ecport
}
