% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\docType{class}
\name{RangeToSubset-class}
\alias{RangeToSubset-class}
\alias{RangeToSubset}
\alias{rangetosubset}
\alias{Rangetosubset}
\title{Range S4 class (range) useful for subsetting with actual values instead of indicators}
\description{
Range S4 class (range) useful for subsetting with actual values instead of indicators
}
\section{Slots}{

\describe{
\item{\code{range}}{numeric vector with min and max value}
}}

\author{
Adriaan Blommaert
}
