% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralPreprocessing.R
\name{spectralNormalization}
\alias{spectralNormalization}
\title{normalization function}
\usage{
spectralNormalization(object, method = "normalize",
  wavelengthRange = r(-Inf, Inf), wavelength = NULL, scaleFunction = "sd",
  meanFunction = NULL)
}
\description{
normalization function
}
\keyword{internal}
