% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralPreprocessing.R
\name{spectralSmoothAndDifferentiate}
\alias{spectralSmoothAndDifferentiate}
\title{internal smoothing and differentiation function}
\usage{
spectralSmoothAndDifferentiate(
  object,
  method = "sg",
  order = 3,
  window = order + 7 - order\%\%2,
  derivative = 0,
  dim = "spectralAxis"
)
}
\description{
internal smoothing and differentiation function
}
\keyword{internal}
