% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralIntegration.R
\name{trapz}
\alias{trapz}
\title{integration via trapezium ruls originally from caTools package, included to avoid dependency on old package
with correction inverse scale, should give still possitive values (taken absolute values )}
\usage{
trapz(x, y)
}
\description{
integration via trapezium ruls originally from caTools package, included to avoid dependency on old package
with correction inverse scale, should give still possitive values (taken absolute values )
}
\keyword{internal}
